fix(cmap): add transparency to the color range around zero
