+++ sdcflows/interfaces/reportlets.py
@@ -7,6 +7,8 @@ from niworkflows.viz.utils import cuts_from_bbox, compose_view
 from nipype.interfaces.base import File, isdefined
 from nipype.interfaces.mixins import reporting
 
+from ..viz.utils import plot_registration, coolwarm_transparent
+
 
 class FieldmapReportletInputSpec(reporting.ReportCapableInputSpec):
     reference = File(exists=True, mandatory=True, desc='input reference')
@@ -33,7 +35,6 @@ class FieldmapReportlet(reporting.ReportCapableInterface):
 
     def _generate_report(self):
         """Generate a reportlet."""
-        import numpy as np
         NIWORKFLOWS_LOG.info('Generating visual report')
 
         refnii = load_img(self.inputs.reference)
@@ -43,7 +44,6 @@ class FieldmapReportlet(reporting.ReportCapableInterface):
         cuts = cuts_from_bbox(contour_nii or mask_nii, cuts=self._n_cuts)
         fmapdata = fmapnii.get_fdata()
         vmax = max(fmapdata.max(), abs(fmapdata.min()))
-        thres = np.percentile(fmapdata[fmapdata > 0], 5)
 
         # Call composer
         compose_view(
@@ -59,67 +59,8 @@ class FieldmapReportlet(reporting.ReportCapableInterface):
                               label='fieldmap (Hz)',
                               contour=contour_nii,
                               compress=False,
-                              plot_params={'cmap': 'coolwarm', 
-                                           'vmax': vmax, 
+                              plot_params={'cmap': coolwarm_transparent(),
+                                           'vmax': vmax,
                                            'vmin': -vmax}),
             out_file=self._out_report
         )
-
-def plot_registration(anat_nii, div_id, plot_params=None,
-                      order=('z', 'x', 'y'), cuts=None,
-                      estimate_brightness=False, label=None, contour=None,
-                      compress='auto'):
-    """
-    Plots the foreground and background views
-    Default order is: axial, coronal, sagittal
-    """
-    from uuid import uuid4
-
-    from lxml import etree
-    from nilearn.plotting import plot_anat
-    from svgutils.transform import SVGFigure
-    from niworkflows.viz.utils import robust_set_limits, extract_svg, SVGNS
-
-    plot_params = plot_params or {}
-
-    # Use default MNI cuts if none defined
-    if cuts is None:
-        raise NotImplementedError  # TODO
-
-    out_files = []
-    if estimate_brightness:
-        plot_params = robust_set_limits(anat_nii.get_data().reshape(-1),
-                                        plot_params)
-
-    # Plot each cut axis
-    for i, mode in enumerate(list(order)):
-        plot_params['display_mode'] = mode
-        plot_params['cut_coords'] = cuts[mode]
-        if i == 0:
-            plot_params['title'] = label
-        else:
-            plot_params['title'] = None
-
-        # Generate nilearn figure
-        display = plot_anat(anat_nii, **plot_params)
-        if contour is not None:
-            display.add_contours(contour, colors='g', levels=[0.5],
-                                 linewidths=0.5)
-
-        svg = extract_svg(display, compress=compress)
-        display.close()
-
-        # Find and replace the figure_1 id.
-        try:
-            xml_data = etree.fromstring(svg)
-        except etree.XMLSyntaxError as e:
-            NIWORKFLOWS_LOG.info(e)
-            return
-        find_text = etree.ETXPath("//{%s}g[@id='figure_1']" % SVGNS)
-        find_text(xml_data)[0].set('id', '%s-%s-%s' % (div_id, mode, uuid4()))
-
-        svg_fig = SVGFigure()
-        svg_fig.root = xml_data
-        out_files.append(svg_fig)
-
-    return out_files
